# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEntry": {
        "description": "One row of the event's transition history.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          }
        },
        "required": [
          "id",
          "action",
          "created_at"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
...
      "EventConfig": {
        "description": "The configurable rules of the ritual, stored as JSON on the Event.",
        "properties": {
          "agent_policy": {
            "default": "allowed",
            "enum": [
              "allowed",
              "forbidden"
            ],
            "title": "Agent Policy",
            "type": "string"
          },
          "auto_score": {
            "default": true,
            "title": "Auto Score",
            "type": "boolean"
          },
          "leaderboard_mode": {
            "default": "best",
            "enum": [
              "best",
              "latest"
            ],
            "title": "Leaderboard Mode",
            "type": "string"
          },
          "registration_open": {
            "default": true,
            "title": "Registration Open",
            "type": "boolean"
          },
          "submission_limit_per_participant": {
            "default": 10,
            "title": "Submission Limit Per Participant",
            "type": "integer"
          },
          "submission_limit_window_hours": {
            "default": 24,
            "title": "Submission Limit Window Hours",
            "type": "integer"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "title": "EventConfig",
        "type": "object"
      },
      "EventConfigUpdate": {
        "description": "Partial update of event configuration. Only set fields are changed.",
        "properties": {
          "agent_policy": {
            "anyOf": [
              {
                "enum": [
                  "allowed",
                  "forbidden"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Policy"
          },
          "auto_score": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Score"
          },
          "leaderboard_mode": {
            "anyOf": [
              {
                "enum": [
                  "best",
                  "latest"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leaderboard Mode"
          },
          "registration_open": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Open"
          },
          "submission_limit_per_participant": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Limit Per Participant"
          },
          "submission_limit_window_hours": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Limit Window Hours"
          },
          "tracks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Track"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracks"
          }
        },
        "title": "EventConfigUpdate",
        "type": "object"
      },
      "EventResponse": {
        "description": "Public view of the event the container is hosting.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/EventConfig"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "state": {
            "enum": [
              "DRAFT",
              "OPEN",
              "FROZEN",
              "FINAL",
              "ARCHIVED"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "state",
          "start",
          "end",
          "config"
        ],
        "title": "EventResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StateTransitionRequest": {
        "description": "Admin request to advance (or reopen) the event's state machine.",
        "properties": {
          "confirm": {
            "default": false,
            "description": "Required for the FROZEN\u2192OPEN reopen, which undoes a closing",
            "title": "Confirm",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the ritual is advancing \u2014 recorded in the audit log",
            "title": "Reason"
          },
          "state": {
            "description": "Target state",
            "enum": [
              "DRAFT",
              "OPEN",
              "FROZEN",
              "FINAL",
              "ARCHIVED"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "StateTransitionRequest",
        "type": "object"
      },
      "StateTransitionResponse": {
        "description": "Result of a successful state transition.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "previous_state": {
            "enum": [
              "DRAFT",
              "OPEN",
              "FROZEN",
              "FINAL",
              "ARCHIVED"
            ],
            "title": "Previous State",
            "type": "string"
          },
          "state": {
            "enum": [
              "DRAFT",
              "OPEN",
              "FROZEN",
              "FINAL",
              "ARCHIVED"
            ],
            "title": "State",
            "type": "string"
          },
          "transitioned_at": {
            "format": "date-time",
            "title": "Transitioned At",
            "type": "string"
          },
          "transitioned_by": {
            "title": "Transitioned By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "previous_state",
          "transitioned_at",
          "transitioned_by"
        ],
        "title": "StateTransitionResponse",
        "type": "object"
      },
...
      "Track": {
        "description": "A submission track within the event.",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "Track",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
...
    "title": "HackRitual",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/event/audit": {
      "get": {
        "operationId": "get_audit_api_admin_event_audit_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response Get Audit Api Admin Event Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit",
        "tags": [
          "event"
        ]
      }
    },
    "/api/admin/event/config": {
      "patch": {
        "operationId": "update_config_api_admin_event_config_patch",
        "parameters": [
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "event"
        ]
      }
    },
    "/api/admin/event/state": {
      "post": {
        "operationId": "transition_state_api_admin_event_state_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateTransitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition State",
        "tags": [
          "event"
        ]
      }
    },
...
  },
...
}